api/requirements.txt
Flask
Flask-CORS
bcrypt
PyJWT
python-dotenv